chore: Combine data sources into a single crate
