--- crates/datasource_bigquery/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "datasource_bigquery"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0"
-tracing = "0.1"
-futures = "0.3.28"
-async-trait = "0.1.68"
-tokio = { version = "1", features = ["full"] }
-async-stream = "0.3.5"
-datafusion = {workspace = true}
-serde_json = "1.0.96"
-serde = { version = "1.0", features = ["derive"] }
-gcp-bigquery-client = "0.16.7"
-bigquery-storage = { git = "https://github.com/glaredb/bigquery-storage", branch = "master" }
-parking_lot = "0.12.1"
-async-channel = "1.8.0"
-metastore = { path = "../metastore" }
-datasource_common = { path = "../datasource_common" }
