--- crates/datasource_debug/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "datasource_debug"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0"
-futures = "0.3.28"
-async-trait = "0.1.68"
-serde = { version = "1.0", features = ["derive"] }
-datafusion = {workspace = true}
-datasource_common = { path = "../datasource_common" }
-metastore = { path = "../metastore" }
