--- crates/datasource_mongodb/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "datasource_mongodb"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-datasource_common = {path = "../datasource_common"}
-thiserror = "1.0"
-tracing = "0.1"
-futures = "0.3.28"
-async-trait = "0.1.68"
-tokio = { version = "1", features = ["full"] }
-async-stream = "0.3.5"
-datafusion = {workspace = true}
-serde = { version = "1.0", features = ["derive"] }
-chrono = "*"
-mongodb = "2.5.0"
-bitvec = "1"
