--- crates/datasource_mysql/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "datasource_mysql"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-metastore = { path = "../metastore" }
-datasource_common = { path = "../datasource_common" }
-thiserror = "1.0"
-tracing = "0.1"
-futures = "0.3.28"
-async-trait = "0.1.68"
-tokio = { version = "1", features = ["full"] }
-async-stream = "0.3.5"
-datafusion = {workspace = true}
-serde = { version = "1.0", features = ["derive"] }
-mysql_async = { version = "0.32.2", default-features = false, features = ["default-rustls"] }
-mysql_common = { version = "0.30.4", features = ["chrono"] }
-chrono = "*"
-rust_decimal = "*"
-openssh = "0.9.9"
