--- crates/datasource_object_store/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "datasource_object_store"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0"
-tracing = "0.1"
-futures = "0.3.28"
-async-trait = "0.1.68"
-datafusion = {workspace = true}
-serde = { version = "1.0", features = ["derive"] }
-object_store = { version = "0.5", features = ["gcp", "aws"] }
-bytes = "1.4.0"
