--- crates/datasource_postgres/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "datasource_postgres"
-version = { workspace = true }
-edition = { workspace = true }
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-metastore = { path = "../metastore" }
-datasource_common = { path = "../datasource_common" }
-thiserror = "1.0"
-tracing = "0.1"
-futures = "0.3.28"
-async-trait = "0.1.68"
-tokio = { version = "1", features = ["full"] }
-datafusion = { workspace = true }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.96"
-tokio-postgres = { version = "0.7.8", features = [
-    "with-uuid-1",
-    "with-serde_json-1",
-    "with-chrono-0_4",
-] }
-uuid = "1.3.3"
-chrono = "0.4.24"
-decimal = { path = "../decimal" }
-rust_decimal = { version = "1.29.0", features = ["db-tokio-postgres"] }
