--- crates/datasource_snowflake/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "datasource_snowflake"
-version = {workspace = true}
-edition = {workspace = true}
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-datasource_common = { path = "../datasource_common" }
-thiserror = "1.0"
-async-trait = "0.1.66"
-snowflake_connector = { path = "../snowflake_connector" }
-datafusion = {workspace = true}
-tokio = { version = "1", features = ["full"] }
-async-stream = "0.3.5"
-futures = "0.3.27"
