+++ crates/datasources/src/lib.rs
@@ -1,0 +1,11 @@
+//! Data source implementations.
+
+pub mod bigquery;
+pub mod debug;
+pub mod mongodb;
+pub mod mysql;
+pub mod object_store;
+pub mod postgres;
+pub mod snowflake;
+
+pub mod errors;
