+++ crates/datasources/src/object_store/gcs.rs
@@ -1,43 +1,40 @@
 use std::sync::Arc;
 
 use datafusion::datasource::TableProvider;
-use object_store::aws::AmazonS3Builder;
+use object_store::gcp::GoogleCloudStorageBuilder;
 use object_store::path::Path as ObjectStorePath;
 use object_store::{ObjectMeta, ObjectStore};
 use serde::{Deserialize, Serialize};
 use tracing::trace;
 
-use crate::csv::CsvTableProvider;
-use crate::errors::Result;
-use crate::parquet::ParquetTableProvider;
-use crate::{file_type_from_path, FileType, TableAccessor};
+use super::csv::CsvTableProvider;
+use super::errors::Result;
+use super::parquet::ParquetTableProvider;
+use super::{file_type_from_path, FileType, TableAccessor};
 
-/// Information needed for accessing an external Parquet file on Amazon S3.
+/// Information needed for accessing an external Parquet file on Google Cloud
+/// Storage.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct S3TableAccess {
-    /// S3 object store region
-    pub region: String,
-    /// S3 object store bucket name
+pub struct GcsTableAccess {
+    /// GCS object store bucket name
     pub bucket_name: String,
-    /// S3 object store access key id
-    pub access_key_id: String,
-    /// S3 object store secret access key
-    pub secret_access_key: String,
-    /// S3 object store table location
+    /// GCS object store service account key
+    pub service_acccount_key_json: String,
+    /// GCS object store table location
     pub location: String,
     pub file_type: Option<FileType>,
 }
 
 #[derive(Debug)]
-pub struct S3Accessor {
-    /// S3 object store access info
+pub struct GcsAccessor {
+    /// GCS object store access info
     pub store: Arc<dyn ObjectStore>,
     /// Meta information for location/object
     pub meta: Arc<ObjectMeta>,
     pub file_type: FileType,
 }
 
-impl TableAccessor for S3Accessor {
+impl TableAccessor for GcsAccessor {
     fn store(&self) -> &Arc<dyn ObjectStore> {
         &self.store
     }
@@ -47,15 +44,13 @@ impl TableAccessor for S3Accessor {
     }
 }
 
-impl S3Accessor {
-    /// Setup accessor for S3
-    pub async fn new(access: S3TableAccess) -> Result<Self> {
+impl GcsAccessor {
+    /// Setup accessor for GCS
+    pub async fn new(access: GcsTableAccess) -> Result<Self> {
         let store = Arc::new(
-            AmazonS3Builder::new()
-                .with_region(access.region)
+            GoogleCloudStorageBuilder::new()
+                .with_service_account_key(access.service_acccount_key_json)
                 .with_bucket_name(access.bucket_name)
-                .with_access_key_id(access.access_key_id)
-                .with_secret_access_key(access.secret_access_key)
                 .build()?,
         );
 
@@ -73,13 +68,11 @@ impl S3Accessor {
         })
     }
 
-    pub async fn validate_table_access(access: S3TableAccess) -> Result<()> {
+    pub async fn validate_table_access(access: GcsTableAccess) -> Result<()> {
         let store = Arc::new(
-            AmazonS3Builder::new()
-                .with_region(access.region)
+            GoogleCloudStorageBuilder::new()
+                .with_service_account_key(access.service_acccount_key_json)
                 .with_bucket_name(access.bucket_name)
-                .with_access_key_id(access.access_key_id)
-                .with_secret_access_key(access.secret_access_key)
                 .build()?,
         );
