+++ crates/datasources/src/snowflake/errors.rs
