+++ crates/sqlexec/src/planner/errors.rs
@@ -56,7 +56,7 @@ pub enum PlanError {
     Internal(String),
 
     #[error(transparent)]
-    DatasourceDebug(#[from] datasource_debug::errors::DebugError),
+    DatasourceDebug(#[from] datasources::debug::errors::DebugError),
 
     #[error(transparent)]
     DatasourceCommon(#[from] datasource_common::errors::DatasourceCommonError),
